requirements.txt
ahk==0.12.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.2
idna==3.1
Jinja2==3.0.1
MarkupSafe==2.0.1
MouseInfo==0.1.3
multidict==5.1.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.3
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dotenv==0.17.1
PyTweening==1.0.3
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
